requirements.txt
Django==5.1.4
django-environ==0.12.0
sqlparse==0.5.3
pillow==11.1.0

requests==2.32.3
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
certifi==2025.1.31

Markdown==3.7

cryptography==44.0.1
cffi==1.17.1
pycparser==2.22

asgiref==3.8.1
tzdata==2024.2

black==24.4.2
django-debug-toolbar==4.3.0

